Automatically reformat code
